iSEEfier: a very convenient way to fire up your iSEE instance

iSEEfier provides a set of functionality to quickly create, inspect, and combine initial configuration objects. These can be conveniently passed to the function call to launch iSEE() in this manner. This currently works with the sets of panels provided by the iSEE and iSEEu packages, but can be extended to accommodate the usage of any custom panel (e.g. from iSEEde, iSEEpathways, or any panel developed independently by the user).

See also

Author

Maintainer: Najla Abassi abassi.nejla96@gmail.com (ORCID)

Authors:

Federico Marini marinif@uni-mainz.de (ORCID)