iSEEinit: Create an initial state of an iSEE instance for gene expression visualization

iSEEinit() defines the initial setup of an iSEE instance, recommending tailored visual elements to effortlessly illustrate the expression of a gene list in a single view.

Usage

iSEEinit(
  sce,
  feature.list,
  reddim.type = "TSNE",
  clusters = colnames(colData(sce))[1],
  groups = colnames(colData(sce))[1],
  markdownboard = FALSE,
  dynamicMarkerTable = FALSE
)

Arguments

sce

SingleCellExperiment object

feature.list

A character vector containing a list of genes

reddim.type

A string vector containing the dimensionality reduction type

clusters

A character string containing the name of the clusters/cell-type/state...(as listed in the colData of the sce)

groups

A character string of the groups/conditions...(as it appears in the colData of the sce)

markdownboard

A logical indicating whether or not to include the MarkdownBoard panel in the initial configuration

dynamicMarkerTable

A logical indicating whether or not the DynamicMarkerTable and linked panels should be included in the initial configuration

Value

A list of "Panel" objects specifying the initial state of iSEE instance

Examples

sce <- scRNAseq::RichardTCellData()
#> snapshotDate(): 2024-03-18
#> loading from cache
sce <- scuttle::logNormCounts(sce)
sce <- scater::runPCA(sce)
sce <- scater::runTSNE(sce)
gene_list <- c("ENSMUSG00000026581", "ENSMUSG00000005087", "ENSMUSG00000015437")
cluster <- "stimulus"
group <- "single cell quality"
initial <- iSEEinit(sce = sce, feature.list = gene_list, clusters = cluster, groups = group)